refactored LayerPicker into separate files
