+++ app/package.json
@@ -75,6 +75,7 @@
     "date-fns": "^2.21.3",
     "distinct-colors": "^3.0.0",
     "dotenv": "^8.6.0",
+    "easy-peasy": "^5.0.4",
     "express": "^4.17.1",
     "fs-extra": "^10.0.0",
     "geojson-vt": "^3.2.1",
