+++ app/src/components/map/LayerPicker/LayerPicker.tsx
@@ -1,54 +1,34 @@
 import React, { useEffect, useContext, useRef, useMemo, useState } from 'react';
-import ImportExportIcon from '@mui/icons-material/ImportExport';
 import { SortableContainer, SortableElement, SortableHandle } from 'react-sortable-hoc';
-import { DataBCLayer } from '../LayerLoaderHelpers/DataBCRenderLayer';
 import { DomEvent } from 'leaflet';
 import { MapRequestContext } from 'contexts/MapRequestsContext';
 /* HelperFiles Parent Layers */
-import { sortArray, getChild, sortObject, updateChild } from './Sorting/SortLayerOrder';
+import { sortArray, sortObject } from './Sorting/SortLayerOrder';
 /* Helper Files Parent Actions */
-import {
-  DialogCloseBtn,
-  getChildAction,
-  toggleDialog,
-  updateParentAction
-} from './LayersActionsHelper/LayersActionsFunctions';
-import { getParentAction } from 'components/map/LayerPicker/LayersActionsHelper/LayersActionsFunctions';
+import { updateParentAction } from './LayersActionsHelper/LayersActionsFunctions';
 import { toolStyles } from '../Tools/Helpers/ToolStyles';
 // MUI
 import {
   Accordion,
   AccordionSummary,
   Button,
-  Checkbox,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
   Box,
   Grid,
-  IconButton,
   List,
   ListItem,
   ListItemButton,
   ListItemIcon,
   ListItemText,
   Popover,
-  Slider,
   Typography
 } from '@mui/material';
-import ColorLens from '@mui/icons-material/ColorLens';
-import ExpandLessIcon from '@mui/icons-material/ExpandLess';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 // MUI Icons
 import LayersIcon from '@mui/icons-material/Layers';
-import SettingsIcon from '@mui/icons-material/Settings';
 import KMLUpload from 'components/map-buddy-components/KMLUpload';
-import { ColorPicker } from 'mui-color';
 // import PopupState, { bindPopover, bindTrigger } from 'material-ui-popup-state';
-import { IndependentLayer } from '../LayerLoaderHelpers/IndependentRenderLayers';
-import { LayersSelector } from './LayerModeSelector';
 import { KMLShapesUpload } from '../../map-buddy-components/KMLShapesUpload';
+import { ChildLayer } from './Components/ChildLayer';
+import { ParentLayer } from './Components/ParentLayer';
 
 export const LayerPicker = React.memo(
   (props: any) => {
@@ -80,24 +60,11 @@ export const LayerPicker = React.memo(
       }
     }, []);
 
-    /**
-     * Function used to print opacity as text
-     * @param value float to indicate an opacity between 0 and 1
-     * @returns string value
-     */
-    const opacityText = (value: number) => {
-      return `${value.toFixed(1)}`;
-    };
-
     const SortableParentLayer = SortableElement(({ parent }) => {
       const onParentLayerAccordionChange = (event: any, expanded: any) => {
         updateParentAction(layersActions, setLayersActions, parent.id, { expanded: expanded });
       };
-      const DragHandle = SortableHandle(() => (
-        <ListItemIcon>
-          <ImportExportIcon />
-        </ListItemIcon>
-      ));
+
       return (
         <ListItem
           key={Math.random()}
@@ -105,122 +72,7 @@ export const LayerPicker = React.memo(
           dense={true}
           ContainerComponent="div"
           style={{ width: '100%', maxWidth: 440 }}>
-          <Accordion
-            id="parent-accordion"
-            expanded={getParentAction(layersActions, parent.id).expanded}
-            onChange={onParentLayerAccordionChange}
-            style={{ width: '100%' }}>
-            <Grid
-              id="accordion-grid"
-              container
-              style={{ marginTop: -10, marginBottom: -10 }}
-              alignItems="center"
-              xs={12}>
-              <Grid id="accordion-summary" item xs={10}>
-                <AccordionSummary>
-                  <Typography variant="subtitle1">{parent.name}</Typography>
-                  <ListItemIcon>
-                    {getParentAction(layersActions, parent.id).expanded ? <ExpandMoreIcon /> : <ExpandLessIcon />}
-                  </ListItemIcon>
-                </AccordionSummary>
-              </Grid>
-              {/* DragHandle */}
-              <Grid id="draghandle" item xs={1}>
-                <DragHandle />
-              </Grid>
-            </Grid>
-            {parent?.children.map((child: any) => (
-              <Grid
-                id={child.id}
-                container
-                style={{ marginBottom: -5, marginTop: -5 }}
-                key={'layer' + child.id + Math.random()}
-                direction="row"
-                alignItems="center">
-                &emsp;
-                <Grid item xs={2} alignContent="center" justifyContent="center">
-                  <Checkbox
-                    id="child-checkbox"
-                    checked={getChild(layers, parent.id, child.id).enabled}
-                    name={child.name}
-                    onChange={() => {
-                      updateChild(layers, setLayers, parent.id, child.id, {
-                        enabled: !getChild(layers, parent.id, child.id).enabled
-                      });
-                    }}
-                  />
-                </Grid>
-                <Grid item xs={6}>
-                  <Typography variant="caption">{child.name}</Typography>
-                </Grid>
-                {/* Settings Dialog Box */}
-                {process.env.REACT_APP_REAL_NODE_ENV === 'development' && (
-                  <LayerModeDialog parent={parent} child={child} />
-                )}
-                {process.env.REACT_APP_REAL_NODE_ENV === 'local' && <LayerModeDialog parent={parent} child={child} />}
-                <Grid item xs={1}>
-                  <IconButton
-                    id="colorpicker-btn"
-                    className={toolClass.toolBtn}
-                    onClick={() =>
-                      toggleDialog(layersActions, setLayersActions, parent, child, { dialog_colorpicker_open: true })
-                    }>
-                    <ColorLens id="color-lens" style={{ color: child.color_code }} />
-                  </IconButton>
-                  <Dialog
-                    id="layer-settings-dialog"
-                    open={getChildAction(layersActions, parent.id, child.id).dialog_colorpicker_open}
-                    onClose={() =>
-                      toggleDialog(layersActions, setLayersActions, parent, child, {
-                        dialog_colorpicker_open: false
-                      })
-                    }>
-                    <DialogTitle>{child.name}</DialogTitle>
-                    {/* Opacity */}
-                    <DialogContent id="layer-opacity" style={{ width: 300 }}>
-                      <Typography id="slider-title" style={{ marginRight: 10 }}>
-                        Opacity
-                      </Typography>
-                      <Slider
-                        id="slider-control"
-                        defaultValue={child.opacity}
-                        onChangeCommitted={(event: any, newOpacity: number | number[]) => {
-                          updateChild(layers, setLayers, parent.id, child.id, {
-                            opacity: newOpacity
-                          });
-                        }}
-                        getAriaValueText={opacityText}
-                        step={0.0001}
-                        min={0.0}
-                        max={1.0}
-                      />
-                    </DialogContent>
-                    {/* Color Picker */}
-                    {child.layer_mode !== 'wms_online' && (
-                      <DialogContent id="layer-colorpicker" style={{ height: 300 }}>
-                        <ColorPicker
-                          defaultValue={child.color_code}
-                          onChange={(color: any) => {
-                            updateChild(layers, setLayers, parent.id, child.id, { color_code: color });
-                          }}
-                        />
-                      </DialogContent>
-                    )}
-                    <DialogActions>
-                      <DialogCloseBtn
-                        parent={parent}
-                        child={child}
-                        fieldsToUpdate={{ dialog_colorpicker_open: false }}
-                      />
-                    </DialogActions>
-                  </Dialog>
-                </Grid>
-                {/* <Grid item xs={2} style={{ position: 'relative' }}>
-                  {child.loaded === 100 ? <DoneIcon /> : <div>{getErrorIcon(timeLeft)}</div>}
-                        </Grid> */}
-              </Grid>
-            ))}
-          </Accordion>
+          <ParentLayer parent={parent} onParentLayerAccordionChange={onParentLayerAccordionChange} />
         </ListItem>
       );
     });
@@ -257,42 +109,6 @@ export const LayerPicker = React.memo(
 
     return (
       <>
-        {layers.map((parent) => (
-          <div key={Math.random()}>
-            {parent.children.map(
-              (child) =>
-                child.enabled &&
-                (child.bcgw_code ? (
-                  <div key={Math.random()}>
-                    <DataBCLayer
-                      opacity={child.opacity}
-                      bcgw_code={child.bcgw_code}
-                      layer_mode={child.layer_mode}
-                      dataBCAcceptsGeometry={child.dataBCAcceptsGeometry}
-                      simplifyPercentage={child.simplifyPercentage}
-                      inputGeo={props.inputGeo}
-                      color_code={child.color_code}
-                      zIndex={parent.zIndex + child.zIndex}
-                    />
-                  </div>
-                ) : (
-                  <div key={Math.random()}>
-                    <IndependentLayer
-                      opacity={child.opacity}
-                      layer_code={child.layer_code}
-                      bcgw_code={child.bcgw_code}
-                      activity_subtype={child.activity_subtype}
-                      poi_type={child.poi_type}
-                      layer_mode={child.layer_mode}
-                      inputGeo={props.inputGeo}
-                      color_code={child.color_code}
-                      zIndex={parent.zIndex + child.zIndex}
-                    />
-                  </div>
-                ))
-            )}
-          </div>
-        ))}
         <Box id="el">
           <Popover
             id={popoverId}
@@ -353,37 +169,3 @@ export const LayerPicker = React.memo(
     return true;
   }
 );
-
-const LayerModeDialog = (props) => {
-  const mapLayersContext = useContext(MapRequestContext);
-  const { layersActions, setLayersActions } = mapLayersContext;
-  const action = getChildAction(layersActions, props.parent.id, props.child.id);
-
-  return (
-    <Grid item xs={2}>
-      <IconButton
-        id="settings-btn"
-        onClick={() =>
-          toggleDialog(layersActions, setLayersActions, props.parent, props.child, { dialog_layerselector_open: true })
-        }>
-        <SettingsIcon />
-      </IconButton>
-      <Dialog
-        id="layermode-settings-dialog"
-        open={action.dialog_layerselector_open}
-        onClose={() =>
-          toggleDialog(layersActions, setLayersActions, props.parent, props.child, { dialog_layerselector_open: false })
-        }>
-        <DialogTitle>{props.child.name}</DialogTitle>
-        <LayersSelector parent={props.parent} child={props.child} />
-        <DialogActions id="close-dialog-action">
-          <DialogCloseBtn
-            parent={props.parent}
-            child={props.child}
-            fieldsToUpdate={{ dialog_layerselector_open: false }}
-          />
-        </DialogActions>
-      </Dialog>
-    </Grid>
-  );
-};
