+++ app/src/components/map/MapContainer.tsx
@@ -31,6 +31,8 @@ import { MapRecordsContext, MapRecordsContextProvider } from 'contexts/MapRecord
 import { Capacitor } from '@capacitor/core';
 import MapRecordsDataGrid from './MapRecordsDataGrid';
 import { MapRequestContext, MapRequestContextProvider } from 'contexts/MapRequestsContext';
+import { DataBCLayer } from './LayerLoaderHelpers/DataBCRenderLayer';
+import { IndependentLayer } from './LayerLoaderHelpers/IndependentRenderLayers';
 
 //Added comment
 
@@ -302,6 +304,7 @@ const MapContainer: React.FC<IMapContainerProps> = (props) => {
 
           <FlyToAndFadeContextProvider>
             <MapRequestContextProvider>
+              <Layers inputGeo={props.geometryState.geometry} />
               <ZoomButtons position="bottomleft" />
               <ScaleControl position="bottomleft" imperial={false} />
 
@@ -332,6 +335,51 @@ const MapContainer: React.FC<IMapContainerProps> = (props) => {
           </FlyToAndFadeContextProvider>
         </ReactLeafletMapContainer>
       </ReactLeafletEditable>
+    </>
+  );
+};
+
+const Layers = (props: any) => {
+  const { layers } = useContext(MapRequestContext);
+
+  return (
+    <>
+      {layers.map((parent) => (
+        <div key={Math.random()}>
+          {parent.children.map(
+            (child) =>
+              child.enabled &&
+              (child.bcgw_code ? (
+                <div key={Math.random()}>
+                  <DataBCLayer
+                    opacity={child.opacity}
+                    bcgw_code={child.bcgw_code}
+                    layer_mode={child.layer_mode}
+                    dataBCAcceptsGeometry={child.dataBCAcceptsGeometry}
+                    simplifyPercentage={child.simplifyPercentage}
+                    inputGeo={props.inputGeo}
+                    color_code={child.color_code}
+                    zIndex={parent.zIndex + child.zIndex}
+                  />
+                </div>
+              ) : (
+                <div key={Math.random()}>
+                  <IndependentLayer
+                    opacity={child.opacity}
+                    layer_code={child.layer_code}
+                    bcgw_code={child.bcgw_code}
+                    activity_subtype={child.activity_subtype}
+                    poi_type={child.poi_type}
+                    layer_mode={child.layer_mode}
+                    inputGeo={props.inputGeo}
+                    color_code={child.color_code}
+                    zIndex={parent.zIndex + child.zIndex}
+                  />
+                </div>
+              ))
+          )}
+        </div>
+      ))}
     </>
   );
 };
